# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wordEditor 论文排版 API",
    "version": "0.1.0",
    "description": "Markdown ↔ docx 双向排版服务：Markdown 经 Pandoc + OOXML 后处理管线产出符合模板版式的 Word 文档，docx 可反向提取为带图片资源的 Markdown 文件夹。\n\n同一套契约由两个引擎实现，产物按命名空间归一后的 XML 逐部件等价：\n- api-node（Fastify，默认）：`pnpm --filter wordeditor-api-node dev`\n- api-python（FastAPI，备用）：`python -m uvicorn app:app --app-dir services/api-python`\n\n两端响应头、状态码、错误体 `{\"detail\": msg}` 与 SSE 帧格式逐字节一致；仅框架自身产生的差异（404/405、非法 JSON 的 500、`python.path` 路径写法、`/` 的 hint 文案）除外。",
    "x-ai": {
      "engines": ["api-node", "api-python"],
      "defaultBaseUrl": "http://localhost:8787",
      "auth": "none（本机服务，无鉴权；对外暴露请自行加反代鉴权）",
      "typicalFlow": [
        "GET /api/templates 取 templateId",
        "POST /api/build/stream（SSE）构建，done 事件给出 downloadUrl",
        "GET downloadUrl 下载 docx"
      ],
      "constraints": {
        "requestBodyLimit": "96MB（上传整个文件夹时按条目 base64 传入）",
        "uploadEntryLimit": "单条目 base64 长度上限 ≈ 27.4MB（原始约 20MB），超出即跳过并在 log 事件里回报",
        "heavyTaskConcurrency": "重任务（构建 / 预览 / 导入）有并发闸门，槽满排队，SSE 里回报 `[排队] …（第 N 位）`；可用 WORDEDITOR_MAX_CONCURRENCY 调整",
        "jobTtl": "任务目录默认 6 小时后回收（WORDEDITOR_JOB_TTL_MINUTES），downloadUrl 随之失效",
        "pathGuard": "/api/file 与 entries[].relPath 均限定在仓库根 / 任务目录内，越界条目被跳过或 404"
      },
      "streamingNotes": "SSE 每 15 秒发一次注释帧 `:ping`；错误以 event:error 帧承载且 HTTP 状态仍为 200，因此必须读流判断成败，不能只看状态码。"
    }
  },
  "servers": [
...
  ],
...
  "paths": {
...
    "/api/templates": {
      "get": {
        "tags": ["templates"],
        "operationId": "templates",
        "summary": "模板注册表",
        "description": "config/templates.json 原文，外加每条 `reference_exists`（参考 docx 是否就位）。构建入参的 templateId 取这里的 id。",
        "x-curl": "curl -s localhost:8787/api/templates | jq '.templates[].id'",
        "responses": {
          "200": {
            "description": "模板列表",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Templates" }
              }
            }
          },
          "404": { "description": "路由未命中（框架级 404）" },
          "500": {
            "description": "config/templates.json 不存在",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        },
        "x-errors": [{ "status": 500, "detail": "config/templates.json 不存在" }]
      }
    },
...
    "/api/build/stream": {
      "post": {
        "tags": ["build"],
        "operationId": "buildStream",
        "summary": "Markdown → docx 构建（SSE 实时进度）",
        "description": "返回 text/event-stream。事件序列：step(prepare) → log* → step(pandoc/structure/ooxml)* → done | error。\n- `step`：`{id: prepare|pandoc|structure|ooxml, status: wait|process|finish, message?}`\n- `log`：`{line, stream: stdout|stderr}`（Pandoc 与后处理的原样输出）\n- `done`：`{jobId, fileName, downloadUrl}`\n- `error`：`{error}`，出现即构建失败\n每 15 秒一条注释帧 `:ping`（防反代 read timeout），非事件，解析器应忽略。失败也以 HTTP 200 + error 帧回报。",
        "x-curl": "curl -Ns localhost:8787/api/build/stream -H 'Content-Type: application/json' -d '{\"templateId\":\"hutb-guanke\",\"markdown\":\"# T\\n\\n正文\"}'",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BuildRequest" } } }
        },
        "responses": {
          "200": {
            "description": "SSE 事件流",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          }
        },
        "x-errors-in-band": true,
        "x-errors": [
          { "event": "error", "when": "templateId 为空", "error": "templateId is required" },
          { "event": "error", "when": "markdown 与 entries 都为空", "error": "markdown 或 entries 必填" },
          { "event": "error", "when": "未安装 Pandoc", "error": "未检测到 Pandoc：brew install pandoc …" },
          { "event": "error", "when": "mdRelPath 不在上传清单里", "error": "mdRelPath 未在上传列表中: …" },
          { "event": "error", "when": "Pandoc 或后处理失败", "error": "build failed: <exit code>" }
        ]
      }
    },
    "/api/build/download": {
      "get": {
        "tags": ["build"],
        "operationId": "buildDownload",
        "summary": "下载构建产物 docx",
        "description": "jobId 来自 done 事件；任务目录超 TTL 后返回 404。文件名为 UTF-8 编码的 Content-Disposition。",
        "x-curl": "curl -NsOJ 'localhost:8787/api/build/download?jobId=<uuid>&fileName=out.docx'",
        "parameters": [
          {
            "name": "jobId",
            "in": "query",
            "required": true,
            "description": "done 事件里的 jobId，须匹配 ^[\\w-]+$",
            "schema": { "type": "string", "pattern": "^[\\w-]+$" }
          },
          {
            "name": "fileName",
            "in": "query",
            "required": false,
            "description": "下载文件名，缺省 export.docx；会被净化为 [\\w.()-空格中文]",
            "schema": { "type": "string", "default": "export.docx" }
          }
        ],
        "responses": {
          "200": {
            "description": "docx 二进制",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "description": "invalid jobId" },
          "404": { "description": "file not found or expired" }
        },
        "x-errors": [
          { "status": 400, "detail": "invalid jobId" },
          { "status": 404, "detail": "file not found or expired" }
        ]
      }
    },
    "/api/preview/styles": {
      "post": {
        "tags": ["build"],
        "operationId": "previewStyles",
        "summary": "用一段 styles.yaml 即时套版（非流式）",
        "description": "把请求里的 stylesYaml 覆盖到模板上，跑完整后处理，同步返回与 done 事件同形的载荷。适合边改 DSL 边看效果。",
        "x-curl": "curl -s localhost:8787/api/preview/styles -H 'Content-Type: application/json' -d '{\"templateId\":\"hutb-guanke\",\"stylesYaml\":\"heading1:\\n  size: 16\\n\"}'",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PreviewStylesRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "产物载荷，用 downloadUrl 下载",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DownloadPayload" } }
            }
          },
          "400": { "description": "templateId is required / stylesYaml is required" },
          "500": { "description": "style preview failed" }
        },
        "x-errors": [
          { "status": 400, "detail": "templateId is required" },
          { "status": 400, "detail": "stylesYaml is required" },
          { "status": 500, "detail": "style preview failed\n<构建输出>" }
        ]
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "description": "所有非流式错误的统一响应体（与 FastAPI HTTPException 同构）",
        "properties": { "detail": { "type": "string" } },
        "required": ["detail"]
      },
...
      "Template": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "hutb-guanke" },
          "name": { "type": "string" },
          "reference_doc": { "type": "string", "description": "仓库根相对的 reference.docx" },
          "reference_exists": { "type": "boolean", "description": "服务端附带" },
          "styles_yaml": { "type": "string" },
          "lua_filter": { "type": "string" },
          "heading_numbering": { "type": "string", "enum": ["gongke", "guanke"] },
          "three_line_tables": { "type": "boolean" },
          "cover_block": { "type": "string" },
          "tail_block": { "type": "string" },
          "note": { "type": "string" }
        },
        "required": ["id", "name", "reference_doc"]
      },
      "Templates": {
        "type": "object",
        "properties": {
          "default_template": { "type": "string" },
          "templates": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Template" }
          }
        },
        "required": ["templates"]
      },
...
      "BuildOptions": {
        "type": "object",
        "additionalProperties": false,
        "description": "版式覆写项，全部可选；见 docs/styles-dsl.md",
        "properties": {
          "noHtmlPipe": { "type": "boolean", "description": "跳过 md→html→docx 两段管道，更快但 HTML 片段支持弱" },
          "noPostprocess": { "type": "boolean", "description": "跳过结构类后处理（document/styles/三线表/逐字表），页眉页脚与密码仍按入参执行" },
          "password": { "type": "string", "description": "Word「修改密码」writeProtection，空则不设置" },
          "headerText": { "type": "string", "description": "页眉文案，支持 <page>/<pages>/{page}/{pages}" },
          "headerAlign": { "type": "string", "enum": ["left", "center", "right", "both", "distribute"], "default": "center" },
          "headerVerticalAlign": { "type": "string", "enum": ["top", "center", "baseline", "bottom", "auto"], "default": "center" },
          "footerText": { "type": "string" },
          "footerAlign": { "type": "string", "enum": ["left", "center", "right", "both", "distribute"], "default": "center" },
          "footerVerticalAlign": { "type": "string", "enum": ["top", "center", "baseline", "bottom", "auto"], "default": "center" }
        }
      },
      "DocxProvenance": {
        "type": "object",
        "additionalProperties": false,
        "description": "写入 docProps 的文档属性",
        "properties": {
          "author": { "type": "string" },
          "remark": { "type": "string", "description": "对应 docProps 的 comments" },
          "title": { "type": "string", "description": "覆盖文档标题" }
        }
      },
      "UploadEntry": {
        "type": "object",
        "description": "上传文件夹里的一个条目（图片、被引用的 .md、bib 等）",
        "properties": {
          "relPath": { "type": "string", "description": "任务目录相对路径，越界条目会被跳过" },
          "contentBase64": { "type": "string", "description": "文件内容 base64" }
        },
        "required": ["relPath", "contentBase64"]
      },
      "BuildRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "两种输入模式：单文件传 markdown；整个文件夹传 entries + mdRelPath。",
        "properties": {
          "templateId": { "type": "string", "description": "GET /api/templates 的 id；hutb-shared 是 hutb-guanke 的别名", "example": "hutb-guanke" },
          "markdown": { "type": "string", "description": "主 Markdown 原文（与 entries 二选一）" },
          "entries": { "type": "array", "items": { "$ref": "#/components/schemas/UploadEntry" } },
          "mdRelPath": { "type": "string", "description": "entries 模式下主 md 的相对路径，如 论文/正文.md" },
          "fileName": { "type": "string", "description": "下载文件名，缺省取模板名 + 标题；会被净化" },
          "options": { "$ref": "#/components/schemas/BuildOptions" },
          "provenance": { "$ref": "#/components/schemas/DocxProvenance" }
        },
        "required": ["templateId"]
      },
      "PreviewStylesRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "templateId": { "type": "string", "example": "hutb-guanke" },
          "stylesYaml": { "type": "string", "description": "临时覆写的 styles.yaml DSL，见 docs/styles-dsl.md" }
        },
        "required": ["templateId", "stylesYaml"]
      },
...
      "DownloadPayload": {
        "type": "object",
        "description": "产物定位载荷",
        "properties": {
          "jobId": { "type": "string", "format": "uuid" },
          "fileName": { "type": "string" },
          "downloadUrl": { "type": "string", "example": "/api/build/download?jobId=…&fileName=….docx" }
        },
        "required": ["jobId", "fileName", "downloadUrl"]
      }
    }
  }
}